blood test analyser/requirements.txt
celery>=5.2.7,<6.0
redis
crewai==0.134.0
importlib-metadata>=6.8.0
opentelemetry-api>=1.24.0
SQLAlchemy
python-dotenv
langchain
openai
